tools/autocaption/requirements.txt
git+https://github.com/huggingface/transformers.git
fastapi
Pillow
torch
accelerate
uvicorn
python-multipart
clip